flask_s3_backend/requirements.txt
Flask
boto3
python-dotenv
flask-cors
supabase